requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow>=9.0.0
markdown==3.5.1
bleach==6.1.0
python-slugify==8.0.1
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
